Move jest integration utils to top level test/utils
