+++ framework/test/jest/integration/specs/modules/chain/components/storage/entities/account/get.spec.js
@@ -15,7 +15,7 @@
 const {
 	Account,
 } = require('../../../../../../../../../../src/modules/chain/components/storage/entities');
-const { PgHelper } = require('../../../../../../../utils/pg-helper');
+const { PgHelper } = require('../../../../../../../../../utils/pg-helper');
 const { constants } = require('../../../../../../../../../utils');
 
 describe('storage.entities.Account.get', () => {
