+++ framework/test/jest/integration/specs/modules/chain/components/storage/entities/round_delegates/summed_round.spec.js
@@ -15,7 +15,7 @@
 const {
 	RoundDelegates,
 } = require('../../../../../../../../../../src/modules/chain/components/storage/entities');
-const { PgHelper } = require('../../../../../../../utils/pg-helper');
+const { PgHelper } = require('../../../../../../../../../utils/pg-helper');
 
 describe('storage.entities.RoundDelegates.summedRound', () => {
 	let pgHelper;
