+++ framework/test/utils/index.js
@@ -18,7 +18,14 @@ const { constants } = require('./constants');
 
 const randomInt = (low, high) => Math.round(Math.random() * (high - low) + low);
 
+const storageUtils = require('./storage');
+const configUtils = require('./configs');
+const chainUtils = require('./chain');
+
 module.exports = {
 	constants,
 	randomInt,
+	storageUtils,
+	configUtils,
+	chainUtils,
 };
